+++ CHANGELOG.md
@@ -1,4 +1,11 @@
 # Change log
 
+## 1.1.0 (2021-12-25)
+- add supports the use of `open` and `close` properties for each style
+- fix codes for methods ansi256() and bgAnsi256()
+- add demo to npm package
+- update package.json
+- update readme
+
 ## 1.0.0 (2021-12-24)
 First release